requirements.txt
aiohttp==2.0.7
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
async-timeout==1.2.1
billiard==3.3.0.23
celery==3.1.25
chardet==3.0.2
colorama==0.3.9
Django==1.11
django-celery==3.2.1
django-debug-toolbar==1.8
djangorestframework==3.6.2
gevent==1.2.1
greenlet==0.4.12
kombu==3.0.37
lxml==3.7.3
multidict==2.1.4
packaging==16.8
psycopg2==2.7.1
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
redis==2.10.5
requests==2.13.0
six==1.10.0
sqlparse==0.2.3
vine==1.1.3
yarl==0.10.1
